+++ examples/minimal-appdir/src/app/layout.tsx
@@ -2,6 +2,11 @@ import "@uploadthing/react/styles.css";
 
 import { Inter } from "next/font/google";
 
+import { NextSSRPlugin } from "@uploadthing/react/next-ssr-plugin";
+import { extractRouterConfig } from "uploadthing/server";
+
+import { uploadRouter } from "~/server/uploadthing";
+
 const inter = Inter({ subsets: ["latin"] });
 
 export default function RootLayout({
@@ -11,7 +16,10 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <NextSSRPlugin routerConfig={extractRouterConfig(uploadRouter)} />
+        {children}
+      </body>
     </html>
   );
 }