+++ packages/uploadthing/src/next-legacy.ts
@@ -4,13 +4,14 @@ import { getStatusCodeFromError, UploadThingError } from "@uploadthing/shared";
 import type { Json } from "@uploadthing/shared";
 
 import { UPLOADTHING_VERSION } from "./constants";
-import { defaultErrorFormatter } from "./internal/error-formatter";
+import { formatError } from "./internal/error-formatter";
 import {
   buildPermissionsInfoHandler,
   buildRequestHandler,
 } from "./internal/handler";
 import type { RouterWithConfig } from "./internal/handler";
-import type { FileRouter, inferErrorShape } from "./internal/types";
+import { incompatibleNodeGuard } from "./internal/incompat-node-guard";
+import type { FileRouter } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
 
@@ -18,15 +19,17 @@ export type { FileRouter } from "./internal/types";
 
 export const createUploadthing = <TErrorShape extends Json>(
   opts?: CreateBuilderOptions<TErrorShape>,
-) => createBuilder<"pages", TErrorShape>(opts);
+) =>
+  createBuilder<
+    { req: NextApiRequest; res: NextApiResponse; event: undefined },
+    TErrorShape
+  >(opts);
 
 export const createNextPageApiHandler = <TRouter extends FileRouter>(
   opts: RouterWithConfig<TRouter>,
 ) => {
-  const requestHandler = buildRequestHandler<TRouter, "pages">(opts);
-  const errorFormatter =
-    opts.router[Object.keys(opts.router)[0]]?._def.errorFormatter ??
-    defaultErrorFormatter;
+  incompatibleNodeGuard();
+  const requestHandler = buildRequestHandler<TRouter>(opts);
 
   const getBuildPerms = buildPermissionsInfoHandler<TRouter>(opts);
 
@@ -46,10 +49,7 @@ export const createNextPageApiHandler = <TRouter extends FileRouter>(
         url,
         json: () =>
           Promise.resolve(
-            JSON.parse(
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-              req.body,
-            ),
+            typeof req.body === "string" ? JSON.parse(req.body) : req.body,
           ),
       }),
       res,
@@ -60,10 +60,7 @@ export const createNextPageApiHandler = <TRouter extends FileRouter>(
     if (response instanceof UploadThingError) {
       res.status(getStatusCodeFromError(response));
       res.setHeader("x-uploadthing-version", UPLOADTHING_VERSION);
-      const formattedError = errorFormatter(
-        response,
-      ) as inferErrorShape<TRouter>;
-      return res.json(formattedError);
+      return res.json(formatError(response, opts.router));
     }
 
     if (response.status !== 200) {