+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @uploadthing/solid
 
+## 5.4.0
+
+### Minor Changes
+
+- [#397](https://github.com/pingdotgg/uploadthing/pull/397)
+  [`7f0c474`](https://github.com/pingdotgg/uploadthing/commit/7f0c4746cf8ed7bac6da3475e0de3a6fe58cf35e)
+  Thanks [@Mr0Bread](https://github.com/Mr0Bread)! - feat: auto upload mode for
+  dropzone component
+
+### Patch Changes
+
+- [#397](https://github.com/pingdotgg/uploadthing/pull/397)
+  [`7f0c474`](https://github.com/pingdotgg/uploadthing/commit/7f0c4746cf8ed7bac6da3475e0de3a6fe58cf35e)
+  Thanks [@Mr0Bread](https://github.com/Mr0Bread)! - fix: remove unused prefix
+  in tw config to ensure correct css output
+
+## 5.3.1
+
+### Patch Changes
+
+- [#384](https://github.com/pingdotgg/uploadthing/pull/384)
+  [`8c01e98`](https://github.com/pingdotgg/uploadthing/commit/8c01e980cf5b695165101b11c741ce6c1b396bdd)
+  Thanks [@Mr0Bread](https://github.com/Mr0Bread)! - fix bug where progress bar
+  styles was not included in the compiled stylesheet
+
+- Updated dependencies
+  [[`ee8533a`](https://github.com/pingdotgg/uploadthing/commit/ee8533a21e82c786537cea5dd7e98fcb71bb5131)]:
+  - @uploadthing/shared@5.2.3
+
+## 5.3.0
+
+### Minor Changes
+
+- [#298](https://github.com/pingdotgg/uploadthing/pull/298)
+  [`5e8016b`](https://github.com/pingdotgg/uploadthing/commit/5e8016b32fc7709dcd855da33dbc2ecf18eac0b5)
+  Thanks [@Mr0Bread](https://github.com/Mr0Bread)! - feat: support theming
+  components
+
+## 5.2.1
+
+### Patch Changes
+
+- [#286](https://github.com/pingdotgg/uploadthing/pull/286)
+  [`a22d74f`](https://github.com/pingdotgg/uploadthing/commit/a22d74fec2c1236b65816a3e95640f9fccb70bca)
+  Thanks [@juliusmarminge](https://github.com/juliusmarminge)! - fix:
+  generateComponents return correct components
+
 ## 5.2.0
 
 ### Minor Changes